Add files via upload
The Order Closing Tool is my simple and efficient go-to tool to close multiple open trades simultaneously. If you are looking for such a tool, this is it! 
Please also note that this is a script and not an EA, however, do not forget to make sure the AutoTrading Enabled Button is enabled, otherwise this script will not work. Additionally, the tool is a script, meaning it will only execute the code after you drag and drop the file onto the chart. To use it again, you will need to drag and drop the file again.

will be adding more features and updates to make the tool more dynamic. If you do not use MetaTrader 4 but use MetaTrader 5, please find the "MetaTrader 5" Folder > "Order Closing Tool (MT5)" for the MT5 version.

Lastly...ENJOY GUYS!